test format date
