+++ src/pages/Main/FlightsList/utils/index.js
@@ -1,0 +1,12 @@
+const formatDate = (dateString) => {
+    const date = new Date(dateString);
+  
+    const formattedDate = date.toLocaleDateString('ru-RU', {
+        month: 'long',
+        day: 'numeric',
+    });
+  
+    return formattedDate;
+};
+
+export default formatDate;
